Поправил баги, вывел номер сотрудника в редактор

+++ src/database/1. create/12. create_worker.sql
@@ -2,7 +2,7 @@ CREATE TABLE "Worker" (
     id       			int PRIMARY KEY,
     first_name			varchar(50),
     last_name           varchar(50),
-	phone 				int,
+	phone 				bigint,
 	position_id			int,
 	username 			varchar(50),
 	password			text,
